+++ src/models/splinter.cpp
@@ -15,8 +15,8 @@ public:
 		if (x.size() != 10 || y.size() != 8) {
 			return false;
 		}
-		float px = x[0], py = x[1], vx = x[2], vy = x[3], rz = x[4], rtz = x[5],
-		      lrps = x[6], rrps = x[7], lvolt = x[8], rvolt = x[9];
+		double px = x[0], py = x[1], vx = x[2], vy = x[3], rz = x[4], rtz = x[5],
+		       lrps = x[6], rrps = x[7], lvolt = x[8], rvolt = x[9];
 		splinter_update(px, py, vx, vy, rz, rtz, lrps, rrps, lvolt, rvolt);
 
 		y[0] = px; y[1] = py; y[2] = vx; y[3] = vy; y[4] = rz; y[5] = rtz; y[6] = lrps; y[7] = rrps;
